app/requirements.txt
aiogram==2.20
matplotlib==3.5.2
Pillow==9.1.1
requests==2.28.0
transformers==4.20.0
timm==0.5.4
torch==1.11.0
torchvision==0.12.0
gunicorn==20.1.0
python-dotenv==0.20.0
sentencepiece==0.1.96
protobuf==3.20.1
SQLAlchemy==1.4.40
beautifulsoup4==4.11.1
sacremoses==0.0.53
